Conflicts:
	src/Cassandra.IntegrationTests/Cassandra.IntegrationTests.csproj
	src/Cassandra.IntegrationTests/Core/BatchStatements.cs
	src/Cassandra.IntegrationTests/Core/CustomTypeTests.cs
	src/Cassandra.IntegrationTests/Core/ParameterizedStatementsTests.cs
	src/Cassandra.IntegrationTests/Core/PreparedStatementsTests.cs
	src/Cassandra.IntegrationTests/Core/TimeUuidSerializationTests.cs
	src/Cassandra.IntegrationTests/Core/TupleTests.cs
	src/Cassandra.IntegrationTests/Core/UdtMappingsTests.cs
	src/Cassandra.IntegrationTests/FoundBugs/CSHARP_42.cs
	src/Cassandra.IntegrationTests/Linq/LinqMappingTests.cs
	src/Cassandra.IntegrationTests/Linq/LinqSessionTests.cs
	src/Cassandra.IntegrationTests/Policies/Tests/LoadBalancingPolicyTests.cs
	src/Cassandra.Tests/ControlConnectionTests.cs
	src/Cassandra.Tests/Mapping/Linq/LinqToCqlUnitTests.cs
	src/Cassandra.Tests/PoliciesUnitTests.cs
	src/Cassandra/Connection.cs

+++ src/Cassandra.IntegrationTests/Core/ConsistencyTests.cs
@@ -29,7 +29,7 @@ using Cassandra.Mapping.Statements;
 
 namespace Cassandra.IntegrationTests.Core
 {
-    [Category("short")]
+    [Category("long")]
     public class ConsistencyTests : TestGlobals
     {
         private ISession _session;
